Lift daemon-related workarounds to NoDaemonGradleExecuter
Due to the new strict command-line checking introduced by
12f1c9f2b53b3721b0d1de294fdda8427340a9f3, this executor now frequently
starts a new daemon, which means it needs to propagate changes like all
other daemon executors
